asss

+++ src/com/douzone/mysite/action/guest/AjaxDeleteAction.java
@@ -1,18 +1,21 @@
 package com.douzone.mysite.action.guest;
 
 import java.io.IOException;
-import java.util.List;
+import java.util.HashMap;
+import java.util.Map;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.douzone.mvc.action.Action;
-import com.douzone.mvc.util.WebUtils;
 import com.douzone.mysite.repository.GuestBookDao;
 import com.douzone.mysite.vo.GuestBookVo;
 
-public class DeleteAction implements Action{
+import net.sf.json.JSONObject;
+
+public class AjaxDeleteAction implements Action {
+
 	@Override
 	public void execute(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
 		request.setCharacterEncoding("utf-8");
@@ -26,9 +29,20 @@ public class DeleteAction implements Action{
 		vo.setPassword(password);
 		
 		
+		boolean passwordCheck = new GuestBookDao().select(vo);
 		
-		new GuestBookDao().delete(vo);
-		//반환으로 문구 결정
-		WebUtils.redirect(request, response, request.getContextPath()+"/guestbook");
+		if(passwordCheck) {
+			new GuestBookDao().delete(vo);
+		}
+			
+		Map<String, Object> map = new HashMap<String, Object>();
+		map.put("result", "success");
+		map.put("data",passwordCheck);
+		
+		response.setContentType("application/json;charset=UTF-8");
+		JSONObject jsonObject = JSONObject.fromObject(map);
+		response.getWriter().print(jsonObject.toString());
+
 	}
+
 }
